chore: fix build properties
