+++ packages/jetbrains-plugin/src/test/kotlin/com/mongodb/jbplugin/meta/BuildInformationTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 class BuildInformationTest {
     @Test
     fun `loads all build information from the resource file`() {
-        assertEquals("0.0.1", BuildInformation.pluginVersion)
-        assertEquals("<none>", BuildInformation.segmentApiKey)
+        assertEquals("", BuildInformation.pluginVersion)
+        assertEquals("", BuildInformation.segmentApiKey)
     }
 }
